refactor: share 531 routine detection
